feat: 안전하게 사진 삭제하도록 수정

+++ sum3years-app/src/main/java/com/sum3years/ui/main/DownloadedScreen.kt
@@ -1,6 +1,7 @@
 package com.sum3years.ui.main
 
 import android.os.Environment
+import android.widget.Toast
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Box
@@ -17,10 +18,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
@@ -29,6 +32,9 @@ import java.io.File
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun DownloadedScreen() {
+    val context = LocalContext.current
+    val forceUpdate = rememberUpdatedState(Unit)
+
     val downloadDir = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
     val images = downloadDir.listFiles { file -> file.isFile && file.extension == "jpg" }
     var deleteTarget: File? by remember { mutableStateOf(null) }
@@ -79,8 +85,17 @@ fun DownloadedScreen() {
             },
             confirmButton = {
                 TextButton(onClick = {
-                    file.delete()
-                    deleteTarget = null
+                    if (file.exists()) {
+                        try {
+                            file.delete()
+                        } finally {
+                            deleteTarget = null
+                        }
+                    } else {
+                        Toast.makeText(context, "존재하지 않는 사진입니다.", Toast.LENGTH_SHORT).show()
+                        deleteTarget = null
+                        forceUpdate.value // 강제 업데이트
+                    }
                 }) {
                     Text("삭제")
                 }
